Start renaming files to not contain the cmd suffix as it's obsolete
