+++ deployer/__main__.py
@@ -1,10 +1,10 @@
 # Import the various subcommands here, they will be automatically
 # registered into the app
-import deployer.commands.cilogon_client_cmd  # noqa: F401
-import deployer.commands.debug_app_and_commands  # noqa: F401
+import deployer.commands.cilogon  # noqa: F401
+import deployer.commands.debug  # noqa: F401
 import deployer.commands.deployer  # noqa: F401
-import deployer.commands.exec.cloud_commands  # noqa: F401
-import deployer.commands.exec.infra_components_commands  # noqa: F401
+import deployer.commands.exec.cloud  # noqa: F401
+import deployer.commands.exec.infra_components  # noqa: F401
 import deployer.commands.generate.billing.cost_table_cmd  # noqa: F401
 import deployer.commands.generate.dedicated_cluster.aws_commands  # noqa: F401
 import deployer.commands.generate.dedicated_cluster.gcp_commands  # noqa: F401
